move app into src
