+++ frontend/src/App.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Switch, Route } from 'react-router-dom';
-import { Admin } from './Router';
-import Dashboard from '../views/Dashboard';
-import Events from '../views/Events';
-import Settings from '../views/Settings';
+import { Admin } from './org/components/Router';
+import Dashboard from './org/views/Dashboard';
+import Events from './org/views/Events';
+import Settings from './org/views/Settings';
 import { OrgInfo, OrgLogo, OrgLinks, Login } from '$modules/SelfSignUp/routes';
 import { ThemeProvider } from '@mui/material/styles';
 import { theme } from '$lib/theme';
