+++ frontend/src/index.js
@@ -1,4 +1,4 @@
-import App from './org/components/App';
+import App from './App';
 import '../stylesheets/index.scss';
 import ReactDOM from 'react-dom';
 import React from 'react';
